# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AllUsersResponse": {
        "additionalProperties": true,
        "properties": {
          "cache_users_needed": {
            "title": "Cache Users Needed",
            "type": "boolean"
          },
          "current_members": {
            "items": {
              "$ref": "#/components/schemas/AllUsersMemberItem"
            },
            "title": "Current Members",
            "type": "array"
          },
          "nginx_route": {
            "title": "Nginx Route",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          }
        },
        "required": [
          "project_name",
          "project_id",
          "current_members",
          "cache_users_needed",
          "nginx_route"
        ],
        "title": "AllUsersResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectsInfoResponse": {
        "additionalProperties": true,
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectInfoItem"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "ProjectsInfoResponse",
        "type": "object"
      },
...
      "TransferBody": {
        "properties": {
          "new_owner_id": {
            "title": "New Owner Id",
            "type": "string"
          }
        },
        "required": [
          "new_owner_id"
        ],
        "title": "TransferBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane for isolated Supabase projects on shared infrastructure.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Supabase Multitenant Projects API",
    "version": "0.13.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/admin/projects-info": {
      "post": {
        "operationId": "get_projects_for_user_api_admin_projects_info_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Projects For User",
        "tags": [
          "project-insights"
        ]
      }
    },
    "/api/admin/projects/{name}/all-users": {
      "get": {
        "description": "Lista todos os usu\u00e1rios dispon\u00edveis para admins.\nComo a API n\u00e3o tem acesso ao cache, retorna uma estrutura\nque o Nginx pode completar ou usa proxy para Nginx.",
        "operationId": "list_all_users_for_admin_api_admin_projects__name__all_users_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllUsersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All Users For Admin",
        "tags": [
          "project-insights"
        ]
      }
    },
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectListItem"
                  },
                  "title": "Response List Projects Api Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{project_name}": {
      "delete": {
        "operationId": "delete_project_api_projects__project_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Step-Up-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Step-Up-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/projects/{project_name}/telemetry/users": {
      "get": {
        "operationId": "get_project_user_telemetry_api_projects__project_name__telemetry_users_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "24h",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUserTelemetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project User Telemetry",
        "tags": [
          "project-insights"
        ]
      }
    },
...
    "/api/projects/{project_name}/transfer": {
      "post": {
        "operationId": "transfer_project_api_projects__project_name__transfer_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_name",
            "required": true,
            "schema": {
              "title": "Project Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Project",
        "tags": [
          "project-insights"
        ]
      }
    },
    "/api/projects/{ref}/execute-function": {
      "post": {
        "operationId": "execute_project_function_api_projects__ref__execute_function_post",
        "parameters": [
          {
            "in": "path",
            "name": "ref",
            "required": true,
            "schema": {
              "title": "Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Execute Project Function Api Projects  Ref  Execute Function Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Project Function",
        "tags": [
          "project-insights"
        ]
      }
    },
...
  }
}